requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.7
pymilvus==2.3.4
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6
mmh3==3.0.0
yt-dlp==2023.12.30
# LangChain packages for ReAct agent
langchain==0.1.17
langchain-openai==0.1.7
langchain-community==0.0.38
langgraph==0.0.53
langchain-core==0.1.52
# Rich for better console output
rich==13.7.0
# YouTube transcript API
youtube-transcript-api==0.6.1 